refactor(schema-form): move isSchemaNode beside AbstractNode to break the core/nodes cycle
