+++ packages/canard/schema-form/src/core/nodes/AbstractNode/index.ts
@@ -1,1 +1,2 @@
 export { AbstractNode } from './AbstractNode';
+export { isSchemaNode } from './filter';
