+++ packages/canard/schema-form/src/core/nodes/BooleanNode/filter.ts
@@ -1,4 +1,4 @@
-import { isSchemaNode } from '../filter';
+import { isSchemaNode } from '../AbstractNode';
 import type { BooleanNode } from './BooleanNode';
 
 /**
