+++ packages/canard/schema-form/src/core/nodes/ObjectNode/filter.ts
@@ -1,4 +1,4 @@
-import { isSchemaNode } from '../filter';
+import { isSchemaNode } from '../AbstractNode';
 import type { ObjectNode } from './ObjectNode';
 
 /**
