+++ packages/canard/schema-form/src/core/nodes/StringNode/filter.ts
@@ -1,4 +1,4 @@
-import { isSchemaNode } from '../filter';
+import { isSchemaNode } from '../AbstractNode';
 import type { StringNode } from './StringNode';
 
 /**
