+++ packages/canard/schema-form/src/core/nodes/filter.ts
@@ -1,4 +1,3 @@
-import { AbstractNode } from './AbstractNode';
 import type { ArrayNode } from './ArrayNode';
 import type { BooleanNode } from './BooleanNode';
 import type { NullNode } from './NullNode';
@@ -8,55 +7,7 @@ import type { StringNode } from './StringNode';
 import type { VirtualNode } from './VirtualNode';
 import type { SchemaNode } from '../types';
 
-/**
- * Type guard to check if a value is a SchemaNode instance.
- *
- * Essential for runtime type checking when working with dynamic node trees,
- * ensuring type safety when processing unknown values from form data or
- * external sources.
- *
- * @param input - The value to check
- * @returns Whether the input is a SchemaNode
- *
- * @example
- * Basic node validation:
- * ```typescript
- * import { isSchemaNode } from '@canard/schema-form';
- *
- * function processNode(value: unknown) {
- *   if (!isSchemaNode(value)) {
- *     throw new Error('Expected a SchemaNode');
- *   }
- *
- *   // TypeScript knows value is SchemaNode
- *   console.log('Node type:', value.type);
- *   console.log('Node path:', value.path);
- *   console.log('Node value:', value.value);
- * }
- * ```
- *
- * @example
- * Safe node traversal:
- * ```typescript
- * function findNodeByPath(root: unknown, targetPath: string): SchemaNode | null {
- *   if (!isSchemaNode(root)) return null;
- *
- *   if (root.path === targetPath) return root;
- *
- *   // Only branch nodes have children
- *   if (isBranchNode(root) && root.children) {
- *     for (const child of root.children) {
- *       const found = findNodeByPath(child.node, targetPath);
- *       if (found) return found;
- *     }
- *   }
- *
- *   return null;
- * }
- * ```
- */
-export const isSchemaNode = (input: any): input is SchemaNode =>
-  input instanceof AbstractNode;
+export { isSchemaNode } from './AbstractNode';
 
 /**
  * Type guard to check if a node is a branch node that can contain children.
